update to use new fips compliant images
